Isolated source of bug in parser - Parser::ident() calls bump()
In application() parsing loop, ty() is repeatedly called. If it
encounters a token of type Ident, then it will *bump* the token and
return whether it is an Uppercase or Lowercase string. This is probably
undesired behavior in the ty() function, since we do not want to bump
this token if it's not a type