Trying to add in evaluation of fixpoint
- There is some bug in evaluation of case arm TmVars during Fix eval.
Not sure if this issue is showing up elsewhere, but it's happening
during term shifting for sure. See system_f/test.sf final example for a
working case. The variable binding is required, otherwise `z` won't be
recognized, and replacing the variable binding of x with y doesn't work
either