Inelegant fix for issue #9 (see description/issue for info)
Why this doesn't work:

During typechecking of a lambda term, we binding a term variable to a
type within the typing context (on a stack). If we hit a type
abstraction term in the body of the abstraction, any type variables we
have bound on the stack are now basically invalidated. Take the
following term as an example:

\X \x: X. \Y \func: X->Y. func x

We first bind the type variable X to TyVar(0), then the term level
variable x is bound on the typing context stack as Var(0): TyVar(0).
When we hit the second type abstraction, we will have a second TyVar(1)
["Y"]. Within the lexical scope of the body of the abstraction 'func',
the universally quantified type 'X' should *actually* have a de Bruijn
index of 1, however we have previously bound Var(0): TyVar(0) on the
stack, so the typechecker will assign the type TyVar(0) [which is the
type variable 'Y'] to the term variable 'x', which is incorrect.

The inelegant solution here is to simply traverse the type context
binding stack when we see a TyAbs term, and increment all TyVar de
Bruijn indices by one before typing the body of the TyAbs, and then
decrement them when we exit.

I would prefer to find a better way of doing this, but I'm not sure how,
at this point.