Experimentation with Product Kinds
Product kinds seem like it might be a useful mechanism with which to
package multiple abstract types into an existential type - e.g. for use
with an "F-ing modules" based system. We appear to be able to correctly
kind- and type-check functors with this method.

Potential issues include how to handle type simplication in the presence
of a necessitated 'projection' type, which indexes a 'product' type.
Given that we can have an existential signature including something
along the lines of this:

`exists X::(*, *) with { insert: X.1 -> X.0 -> X.0 }`

where X is a type variable representing an existentially abstract Product type with Kind (*,
*). We need to be able to now handle type-simplication of form
Type::Projection(Type::Var(_), _), or potentially even deeper-nested
type variable.